Move Event type to protocol crate.
