+++ client/src/event_loop.rs
@@ -7,105 +7,18 @@ use tokio::sync::mpsc;
 
 use mpc_protocol::{
     channel::decrypt_server_channel, decode, hex, snow::Builder,
-    Encoding, HandshakeMessage, MeetingState, OpaqueMessage,
-    ProtocolState, RequestMessage, ResponseMessage, SealedEnvelope,
-    ServerMessage, SessionId, SessionState, TransparentMessage,
+    Encoding, HandshakeMessage, OpaqueMessage, ProtocolState,
+    RequestMessage, ResponseMessage, SealedEnvelope, ServerMessage,
+    SessionId, TransparentMessage,
 };
 
 use super::{decrypt_peer_channel, Peers, Server};
 use crate::{ClientOptions, Error, Result};
 
+pub use mpc_protocol::{Event, JsonMessage};
+
 /// Stream of events emitted by an event loop.
 pub type EventStream = BoxStream<'static, Result<Event>>;
-
-/// Events dispatched by the event loop stream.
-#[derive(Debug)]
-pub enum Event {
-    /// Event dispatched when a handshake with the server
-    /// is completed.
-    ServerConnected {
-        /// Public key of the server.
-        server_key: Vec<u8>,
-    },
-    /// Event dispatched when a handshake with a peer
-    /// has been completed.
-    PeerConnected {
-        /// Public key of the peer.
-        peer_key: Vec<u8>,
-    },
-    /// Binary message received from a peer.
-    BinaryMessage {
-        /// Public key of the peer.
-        peer_key: Vec<u8>,
-        /// Message buffer.
-        message: Vec<u8>,
-        /// Session identifier.
-        session_id: Option<SessionId>,
-    },
-    /// JSON message received from a peer.
-    JsonMessage {
-        /// Public key of the peer.
-        peer_key: Vec<u8>,
-        /// JSON message.
-        message: JsonMessage,
-        /// Session identifier.
-        session_id: Option<SessionId>,
-    },
-
-    /// Event dispatched when a meeting has been created.
-    MeetingCreated(MeetingState),
-
-    /// Event dispatched when a meeting is ready.
-    ///
-    /// A meeting is ready when the limit for the meeting point
-    /// has been reached.
-    MeetingReady(MeetingState),
-
-    /// Event dispatched when a session has been created.
-    SessionCreated(SessionState),
-
-    /// Event dispatched when a session is ready.
-    ///
-    /// A session is ready when all participants
-    /// have completed the server handshake.
-    ///
-    /// Peers can now handshake with each other.
-    SessionReady(SessionState),
-
-    /// Event dispatched when a session is active.
-    ///
-    /// A session is active when all the participants
-    /// have connected to each other.
-    SessionActive(SessionState),
-
-    /// Event dispatched when a session timed out waiting
-    /// for all the participants.
-    SessionTimeout(SessionId),
-
-    /// Event dispatched when a session has been finished.
-    ///
-    /// A session can only be finished when the session owner
-    /// explicitly closes the session.
-    SessionFinished(SessionId),
-
-    /// Event dispatched when the socket is closed.
-    Close,
-}
-
-/// JSON message received from a peer.
-#[derive(Debug)]
-pub struct JsonMessage {
-    contents: Vec<u8>,
-}
-
-impl JsonMessage {
-    /// Deserialize this message.
-    pub fn deserialize<'a, T: serde::de::Deserialize<'a>>(
-        &'a self,
-    ) -> Result<T> {
-        Ok(serde_json::from_slice::<T>(&self.contents)?)
-    }
-}
 
 /// Internal message used to communicate between
 /// the client and event loop.
@@ -399,7 +312,7 @@ where
                 }),
                 Encoding::Json => Ok(Event::JsonMessage {
                     peer_key: public_key.as_ref().to_vec(),
-                    message: JsonMessage { contents },
+                    message: contents.into(),
                     session_id,
                 }),
             }
